+++ release_notes.md
@@ -19,6 +19,8 @@ This section lists features in master, available by [AppVeyor](https://ci.appvey
  --- We added a method to clear the pipeline cache, particularly for testing where you want to test configuration scenarios
  - Added ability to persist RabbitMQ messages
  - Added subscription to blocked/unblocked RMQ channel events. A warning log is created when a channel becomes blocked and an info log is generated when the channel becomes unblocked.
+ - An option now exits on a connection to not declare infrastructure because it is declared elsewhere but just validate it
+ - On Connection, renamed VisibilityTimeout to LockTimeout to more generically describe its purpose
 
 ## Release 8.1.1399 ##
  - Update nuget libs